requirements.txt
chromadb==0.6.3
anthropic>=0.84.0
python-dotenv==1.0.1
psutil==6.1.1
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
gunicorn>=22.0.0
